Rename breadcrumbs and make arrows thicker
